Code from starter-gatsby-blog
This commit contains all the code from starter-gatsby-blog up to commit
f4a10dadabfcae4d8a5cbfbc1cf9e59d49811f4a, squashed into a single commit.

I did this rather than forking because this repo is not really a fork.
It builds on that repo but is not intended to be the same thing.

You can see the starter-gatsby-blog repo here:
https://github.com/contentful/starter-gatsby-blog